feat: add code

+++ entity/items.go
@@ -6,7 +6,7 @@ type MenuItem struct {
 	ID          string         `json:"id" db:"id"`
 	Name        string         `json:"name" db:"name"`
 	Description string         `json:"description" db:"description"`
-	Price       float64        `json:"price" db:"price"`
+	Price       int            `json:"price" db:"price"`
 	Tags        pq.StringArray `json:"tags" db:"tags"`
 	Picture     string         `json:"picture" db:"picture"`
 }
